Fix build error or warning of ISP_SPI, ISP_HID, ISP_DFU and USBD sample code on Keil compiler 6.

+++ SampleCode/StdDriver/USBD_Audio_NAU8822/usbd_audio.c
@@ -1077,7 +1077,7 @@ void AdjFreq(void)
     if((i32PreFlag != g_i32AdjFlag) || (i32Cnt++ > 40000))
     {
         i32PreFlag = g_i32AdjFlag;
-        printf("%d %d %d %d %x\n", g_i32AdjFlag, u32Size, g_usbd_PlayVolumeL, g_usbd_RecVolumeL, FAUDIOCFG);
+        printf("%d %d %d %d %x\n", g_i32AdjFlag, u32Size, g_usbd_PlayVolumeL, g_usbd_RecVolumeL, (uint32_t)FAUDIOCFG);
         i32Cnt = 0;
     }
